#1069 refactor
