+++ src/ch09_person_lesson/_ref/ch09_path.py
@@ -12,7 +12,6 @@ def create_moments_dir_path(moment_mstr_dir: str) -> str:
 
 def create_moment_dir_path(moment_mstr_dir: str, moment_lasso: LassoUnit) -> str:
     """Returns path: moment_mstr_dir\\moments\\moment_rope"""
-    # TODO create create_moments_dir_path function and replace all "create_moments_dir_path(moment_mstr_dir)" with it
     moments_dir = create_moments_dir_path(moment_mstr_dir)
     return create_path(moments_dir, moment_lasso.make_path())
